+++ src/main/java/ru/yandex/practicum/filmorate/dao/LikeDao.java
@@ -7,7 +7,7 @@ import ru.yandex.practicum.filmorate.model.Like;
  */
 public interface LikeDao {
 
-    Like addLike(int id, int userId);
+    Like addLike(long id, long userId);
 
-    void removeLike(int id, int userId);
+    void removeLike(long id, long userId);
 }
