+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -8,11 +8,11 @@ public interface FilmStorage {
 
     Film addFilm(Film film);
 
-    void removeFilm(int id);
+    void removeFilm(long id);
 
     Film updateFilm(Film updateFilm);
 
     Collection<Film> getAllFilms();
 
-    Film findFilmById(int id);
+    Film findFilmById(long id);
 }
